+++ spec/spec_helper.rb
@@ -4,3 +4,4 @@ $LOAD_PATH.unshift File.expand_path('../lib', __dir__)
 require 'plagscan'
 
 require 'webmock/rspec'
+require 'tempfile'