+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Added support for `token` API
+- Added support for some `documents` APIs (create, check and retrieve)
 
 ## [0.0.1](releases/tag/v0.0.1) - 2019-08-13
 ### Added